Create database package. Move database files to database package.
